# docs/openapi.yaml
openapi: 3.0.0
info:
  title: matrix2acrobits
  version: 1.0.0
  description: API proxy to adapt Acrobits Cloud Softphone to Matrix.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/client/fetch_messages:
    post:
      summary: Fetch Messages
      operationId: fetchMessages
      description: |
        Checks for new incoming messages by performing a Matrix /sync on behalf of the user.
        Authentication is handled by the Application Service backend; the `password` field is ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  description: The full Matrix ID (e.g., @user:server.com) of the user to impersonate.
                password:
                  type: string
                  description: This field is ignored. Kept for client compatibility.
                last_id:
                  type: string
                  description: The 'since' token from the last sync.
                last_sent_id:
                  type: string
                  description: The sms_id of the last sent message.
                device:
                  type: string
                  description: Device identifier (e.g., 'ACROBITS').
      responses:
        '200':
          description: Successful sync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchMessagesResponse'
        '401':
          description: Authentication failed (e.g., user not in AS namespace).
  
  /api/client/send_message:
    post:
      summary: Send Message
      operationId: sendMessage
      description: |
        Sends an outgoing message on behalf of a user.

        The `from` field can be:
        - A full Matrix ID (e.g., @user:server.com) - used directly
        - A phone number (e.g., +1234567890) - resolved to Matrix ID via SMS-to-Matrix mapping if available

        If a phone number is provided and a mapping exists, the message is sent on behalf of the mapped Matrix user.
        If a phone number is provided but no mapping exists, the phone number is used as the sender ID.

        Authentication is handled by the Application Service backend; the `password` field is ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - sms_to
                - sms_body
              properties:
                from:
                  type: string
                  description: The sender - either a full Matrix ID (e.g., @user:server.com) or a phone number (e.g., +1234567890).
                password:
                  type: string
                  description: This field is ignored. Kept for client compatibility.
                sms_to:
                  type: string
                  description: Recipient Matrix ID, Alias, or mapped phone number.
                sms_body:
                  type: string
                  description: The message content.
                content_type:
                  type: string
                  default: text/plain
                disposition_notification:
                  type: string
                  description: Opaque string for read receipts.
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sms_id:
                    type: string
                    description: The Matrix event ID of the sent message.
        '400':
          description: Invalid request or recipient.
        '401':
          description: Authentication failed (e.g., user not in AS namespace).

  /api/client/push_token_report:
    post:
      summary: Report Push Token
      operationId: pushTokenReport
      description: |
        Receives and stores device push tokens from Acrobits clients.
        The tokens are persisted in a SQLite database for later use in push notification delivery.
        This endpoint follows the Acrobits Push Token Reporter API specification for POST JSON requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenReportRequest'
      responses:
        '200':
          description: Push token accepted and stored
          content:
            application/json:
              schema:
                type: object
                description: Empty JSON object response
        '400':
          description: Invalid request payload (e.g., missing selector).
        '500':
          description: Server error (e.g., database unavailable).
      example:
        request:
          selector: "12869E0E6E553673C54F29105A0647204C416A2A:7C3A0D14"
          token_msgs: "QVBBOTFiRzlhcVd2bW54bllCWldHOWh4dnRrZ3pUWFNvcGZpdWZ6bWM2dFAzS2J"
          appid_msgs: "com.cloudsoftphone.app"
          token_calls: "Udl99X2JFP1bWwS5gR/wGeLE1hmAB2CMpr1Ej0wxkrY="
          appid_calls: "com.cloudsoftphone.app.pushkit"
        response:
          {}

  /api/internal/map_sms_to_matrix:
    post:
      summary: Register or update an SMS-to-Matrix mapping
      description: |
        Creates or overwrites the mapping between an SMS number (or other identifier) and
        a Matrix room. Requires the `X-Super-Admin-Token` header.
      parameters:
        - in: header
          name: X-Super-Admin-Token
          schema:
            type: string
          required: true
          description: The Application Service token (as_token).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MappingRequest'
      responses:
        '200':
          description: Mapping acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MappingResponse'
        '400':
          description: Invalid request payload.
        '401':
          description: Invalid admin token.
    get:
      summary: Lookup an existing mapping
      description: |
        Returns the stored Matrix handler for a previously mapped identifier. Requires the
        `X-Super-Admin-Token` header.
      parameters:
        - in: header
          name: X-Super-Admin-Token
          schema:
            type: string
          required: true
          description: The Application Service token (as_token).
        - in: query
          name: sms_number
          schema:
            type: string
          required: false
          description: The SMS number or identifier previously mapped. If omitted, the endpoint returns the full list of mappings.
      responses:
        '200':
          description: Mapping found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MappingResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/MappingResponse'
        '401':
          description: Invalid admin token.
        '404':
          description: No mapping exists for the requested identifier.

components:
  schemas:
    Message:
      type: object
      properties:
        sms_id:
          type: string
          description: Unique identifier (Matrix Event ID).
        sending_date:
          type: string
          format: date-time
          description: RFC 3339 timestamp.
        sender:
          type: string
          description: Sender address/ID.
        recipient:
          type: string
          description: Recipient address/ID (Room ID).
        sms_text:
          type: string
          description: Message body.
        content_type:
          type: string
          description: MIME type.
        stream_id:
          type: string
          description: Identifier for the conversation stream (Room ID).
    FetchMessagesResponse:
      type: object
      properties:
        date:
          type: string
          format: date-time
          description: Current server time in RFC 3339.
        received_smss:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        sent_smss:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    MappingRequest:
      type: object
      required:
        - sms_number
        - room_id
      properties:
        sms_number:
          type: string
          description: SMS number or bridged identifier to map.
        matrix_id:
          type: string
          description: Optional Matrix user to use when auto-creating DMs.
        room_id:
          type: string
          description: Matrix room ID where the mapped identifier should route messages.
        user_name:
          type: string
          description: Optional human-friendly name to expose for this mapping (e.g., contact name).
    MappingResponse:
      type: object
      properties:
        sms_number:
          type: string
        matrix_id:
          type: string
        room_id:
          type: string
        user_name:
          type: string
          description: Optional human-friendly name associated with the mapping.
        updated_at:
          type: string
          format: date-time
          description: When the mapping was last written.
    PushTokenReportRequest:
      type: object
      required:
        - selector
      properties:
        selector:
          type: string
          description: |
            SHA1 hash with suffix that uniquely identifies the account.
            Used to reference the account in push notification delivery requests.
            Example: 12869E0E6E553673C54F29105A0647204C416A2A:7C3A0D14
        token_msgs:
          type: string
          description: |
            Base64-encoded push token for regular notifications (APNs remote-notification).
            Required for iOS push notifications.
        appid_msgs:
          type: string
          description: |
            Apple application ID for regular notifications.
            Used in conjunction with token_msgs.
        token_calls:
          type: string
          description: |
            Base64-encoded push token for VoIP/incoming call notifications (APNs pushkit).
            Required for iOS incoming call notifications.
        appid_calls:
          type: string
          description: |
            Apple application ID for incoming call notifications.
            Used in conjunction with token_calls.